% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel_numerical.R
\name{bipartite_closest_pair}
\alias{bipartite_closest_pair}
\title{Find the closest pair of points, one in one class and the other in another class.}
\usage{
bipartite_closest_pair(coords_1, coords_2)
}
\description{
Find the closest pair of points, one in one class and the other in another class.
}
