% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel_numerical.R
\name{find_contiguous_region}
\alias{find_contiguous_region}
\title{Find a contiguous subset of points.}
\usage{
find_contiguous_region(all_points, good_idx, root_idx)
}
\arguments{
\item{all_points}{Data used.}

\item{good_idx}{The "good" points are `all_points[good_idx, ]`.}

\item{root_idx}{The indices returned include this number.}
}
\description{
Find a contiguous subset of points.
}
\details{
Given a set of "good" points and a "root", find a 
group of good points that contains the root and doesn't span gaps
with nearby "bad" points.
}
