% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel.R
\name{fit_branches}
\alias{fit_branches}
\title{Optimize individual branch models given branch assignments.}
\usage{
fit_branches(branchmodel, ...)
}
\description{
This function updates `@dist_df` and `@models`.
}
