% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel.R
\name{fit_branchmodel}
\alias{fit_branchmodel}
\title{Fit a "Y" shape to data}
\usage{
fit_branchmodel(raw_data, max_iter = 20, tol = 0.001)
}
\arguments{
\item{raw_data}{Dataframe with numeric columns.}

\item{max_iter}{Default 20.}

\item{tol}{Stops once less than 100*tol percent of points are reclassified.}
}
\description{
Fit a "Y" shape to data
}
\details{
This function imposes a "Y" shape on data in a (preferably 2D) space.
It represents the shape as three principal curves (from `princurve`), which each point
hard-assigned to one curve. The internal methods iterate
between reassigning the data to the nearest branch and adjusting the branches,
with a heuristic to make the curves roughly meet in the center.
}
