% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel.R
\name{fit_branchmodel_internal}
\alias{fit_branchmodel_internal}
\title{Internal branchmodel function that performs fitting.}
\usage{
fit_branchmodel_internal(branchmodel, max_iter = 20, tol = 1e-04)
}
\description{
Internal branchmodel function that performs fitting.
}
