% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel_numerical.R
\name{get_connected_component}
\alias{get_connected_component}
\title{Find a connected component of a directed graph.}
\usage{
get_connected_component(neighbors, start_idx)
}
\arguments{
\item{neighbors}{neighbors directly reachable from each vertex, so there is an edge from `i` to `neighbors[i, j]`.
(This is built for knn graphs, hence the weird representation.)}

\item{start_idx}{Algorithm starts here and searches outward.}
}
\description{
Find a connected component of a directed graph.
}
