% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel.R
\name{get_issues}
\alias{get_issues}
\title{Check whether a branchmodel object is valid.}
\usage{
get_issues(object, ...)
}
\value{
Returns empty string if everything is fine. Otherwise, some other length-1 character.
}
\description{
Check whether a branchmodel object is valid.
}
