% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel.R
\name{get_simple_sq_distances}
\alias{get_simple_sq_distances}
\title{For each point, fills in the distance to each line segment.}
\usage{
get_simple_sq_distances(branchmodel, ...)
}
\description{
For each point, fills in the distance to each line segment.
}
