% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel_basics.R
\name{plot_branchmodel}
\alias{plot_branchmodel}
\title{Plot the data and the branchmodel object on a scatterplot (returned ggplot2 plot).}
\usage{
plot_branchmodel(branchmodel, main = "")
}
\description{
Currently just plots the first two variables. 
Eventually might switch to plotting a projection onto a cleverly chosen plane.
}
