% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel_numerical.R
\name{princurve_truncate}
\alias{princurve_truncate}
\title{Remove the last few points from a principal curve.}
\usage{
princurve_truncate(pc, n_remove, ...)
}
\description{
The tricky part is the tag, which is not ordered the same way as the others.
It's like the output of order(<positions along curve>) in that `pc$s[pc$tag, ]` is smooth.
Since it indexes `pc$s`, if I remove `pc$s[45,]`, need to make sure `!any(pc$tag==45)`.
}
