% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchmodel.R
\name{relocate_center}
\alias{relocate_center}
\title{Update the branch-point.}
\usage{
relocate_center(branchmodel, iter, previous_max_ambiguity)
}
\description{
Update the branch-point.
}
\details{
This function updates `@center`.
}
